+++ lldb/source/Plugins/Protocol/MCP/Tool.h
@@ -11,35 +11,15 @@
 
 #include "lldb/Core/Debugger.h"
 #include "lldb/Protocol/MCP/Protocol.h"
+#include "lldb/Protocol/MCP/Tool.h"
 #include "llvm/Support/JSON.h"
 #include <string>
 
 namespace lldb_private::mcp {
 
-class Tool {
+class CommandTool : public lldb_protocol::mcp::Tool {
 public:
-  Tool(std::string name, std::string description);
-  virtual ~Tool() = default;
-
-  virtual llvm::Expected<lldb_protocol::mcp::TextResult>
-  Call(const lldb_protocol::mcp::ToolArguments &args) = 0;
-
-  virtual std::optional<llvm::json::Value> GetSchema() const {
-    return llvm::json::Object{{"type", "object"}};
-  }
-
-  lldb_protocol::mcp::ToolDefinition GetDefinition() const;
-
-  const std::string &GetName() { return m_name; }
-
-private:
-  std::string m_name;
-  std::string m_description;
-};
-
-class CommandTool : public mcp::Tool {
-public:
-  using mcp::Tool::Tool;
+  using lldb_protocol::mcp::Tool::Tool;
   ~CommandTool() = default;
 
   virtual llvm::Expected<lldb_protocol::mcp::TextResult>
