+++ lldb/source/Protocol/MCP/CMakeLists.txt
@@ -1,8 +1,11 @@
 add_lldb_library(lldbProtocolMCP NO_PLUGIN_DEPENDENCIES
+  MCPError.cpp
   Protocol.cpp
+  Tool.cpp
 
   LINK_COMPONENTS
     Support
   LINK_LIBS
     lldbUtility
 )
+
